Add refreshing tokens
