+++ tsconfig.app.json
@@ -25,7 +25,6 @@
     /* Aliases */
     "baseUrl": "./src",
     "paths": {
-      "@apolloclient/*": ["apolloclient/*"],
       "@services/*": ["services/*"],
       "@pages/*": ["pages/*"],
       "@components/*": ["components/*"],
